Attempt at mp3 music

+++ qcommon/common.c
@@ -986,7 +986,7 @@ void *SZ_GetSpace (sizebuf_t *buf, size_t length)
 	return data;
 }
 
-void SZ_Write (sizebuf_t *buf, void *data, int length)
+void SZ_Write (sizebuf_t *buf, void *data, size_t length)
 {
 	memcpy (SZ_GetSpace(buf,length),data,length);		
 }
